+++ tests/common/mod.rs
@@ -1,10 +1,13 @@
 // Copyright (C) 2020 Daniel Mueller <deso@posteo.net>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+use std::fmt::Debug;
+
 use hyper::Body;
 use hyper::body::to_bytes;
 use hyper::Client as HttpClient;
 use hyper::http::request::Builder as HttpRequestBuilder;
+use hyper::Error as HyperError;
 use hyper::Request;
 use hyper_tls::HttpsConnector;
 
@@ -13,6 +16,13 @@ use http_endpoint::Endpoint;
 use url::Url;
 
 const HTTP_BIN_BASE_URL: &str = "https://httpbin.org/";
+
+
+#[derive(Debug)]
+pub enum Error<E> {
+  EndpointError(E),
+  HyperError(HyperError),
+}
 
 
 fn request<E>(input: &E::Input) -> Result<Request<Body>, E::Error>
@@ -31,15 +41,18 @@ where
   Ok(request)
 }
 
-pub async fn issue<E>(input: &E::Input) -> Result<E::Output, E::Error>
+pub async fn issue<E>(input: &E::Input) -> Result<E::Output, Error<E::Error>>
 where
   E: Endpoint,
 {
   let client = HttpClient::builder().build(HttpsConnector::new());
-  let request = request::<E>(input)?;
-  let result = client.request(request).await?;
+  let request = request::<E>(input).map_err(Error::EndpointError)?;
+  let result = client.request(request).await.map_err(Error::HyperError)?;
   let status = result.status();
-  let bytes = to_bytes(result.into_body()).await?;
+  let bytes = to_bytes(result.into_body())
+    .await
+    .map_err(Error::HyperError)?;
   let body = bytes.as_ref();
-  E::evaluate(status, body)
+  let output = E::evaluate(status, body).map_err(Error::EndpointError)?;
+  Ok(output)
 }